+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nx-electron",
-  "version": "10.0.0-beta.0",
+  "version": "10.0.0-beta.1",
   "description": "Electron Plugin for Nx",
   "main": "index.js",
   "types": "index.d.ts",
