+++ src/builders/package/package.impl.ts
@@ -1,202 +1,193 @@
-// import { BuilderContext, createBuilder, BuilderOutput } from '@angular-devkit/architect';
-// import { JsonObject } from '@angular-devkit/core';
+import { BuilderContext, BuilderOutput, createBuilder } from '@angular-devkit/architect';
+import { JsonObject } from '@angular-devkit/core';
 
-// import { serialHooks } from 'electron-packager/src/hooks';
-// import { Options as ElectronPackagerOptions } from 'electron-packager';
-// import electronPackager from 'electron-packager';
+import { build, Configuration, PublishOptions, Platform, Arch, createTargets, FileSet, CliOptions } from 'electron-builder';
+import { writeFile, statSync, readFileSync } from 'fs';
+import { join, resolve } from 'path';
+import { promisify } from 'util';
 
-// import { join } from 'path';
-// import { sync as removeSync } from 'rimraf';
-// import { writeFile, readFile, readFileSync, statSync, readdirSync } from 'fs';
-// import { promisify } from 'util';
+import { getSourceRoot } from '../../utils/workspace';
+import { normalizePackgingOptions } from '../../utils/normalize';
 
-// import { getSourceRoot } from '../../utils/workspace';
-// import { normalizePackagingOptions } from '../../utils/normalize';
+import { Observable, from, of } from 'rxjs';
+import { map, concatMap, catchError } from 'rxjs/operators';
+import { platform } from 'os';
 
-// import { Observable, from, of } from 'rxjs';
-// import { map, concatMap, catchError } from 'rxjs/operators';
+import stripJsonComments from 'strip-json-comments';
 
-// import stripJsonComments from 'strip-json-comments';
+try {
+  require('dotenv').config();
+} catch (e) {}
 
-// try {
-//   require('dotenv').config();
-// } catch (e) {}
+const writeFileAsync = (path: string, data: string) => promisify(writeFile)(path, data, { encoding: 'utf8' });
 
-// const writeFileAsync = (path: string, data: string) => promisify(writeFile)(path, data);
-// const readFileAsync = (path: string) => promisify(readFile)(path);
+export interface PackageElectronBuilderOptions extends Configuration {
+  name: string;
+  frontendProject: string;
+  platform: string | string[];
+  arch: string;
+  root: string;
+  prepackageOnly: boolean;
+  sourcePath: string;
+  outputPath: string;
+  publishPolicy?: PublishOptions["publish"];
+}
 
-// export interface PackageElectronBuilderOptions extends ElectronPackagerOptions {
-//   name: string;
-//   frontendProject: string;
-//   out?: string;
-// }
+export interface PackageElectronBuilderOutput extends BuilderOutput {
+  target?: any;
+  outputPath: string | string[];
+}
 
-// export interface PackageElectronBuilderOutput extends BuilderOutput {
-//   target?: any;
-//   outputPath: string | string[];
-// }
+export default createBuilder<JsonObject & PackageElectronBuilderOptions>(run);
 
-// export default createBuilder<JsonObject & PackageElectronBuilderOptions>(run);
+function run(rawOptions: JsonObject & PackageElectronBuilderOptions, context: BuilderContext): Observable<PackageElectronBuilderOutput> { 
+  return from(getSourceRoot(context)).pipe(
+    map(sourceRoot =>
+      normalizePackgingOptions(rawOptions, context.workspaceRoot, sourceRoot)
+    ),
+    map(options => 
+      mergePresetOptions(options)
+    ),
+    map(options => 
+      addMissingDefaultOptions(options)
+    ),
+    concatMap(async (options) => {
+      await beforeBuild(options.root, options.sourcePath, options.name);
 
-// function run(options: JsonObject & PackageElectronBuilderOptions, context: BuilderContext): Observable<PackageElectronBuilderOutput> {
-//   return from(getSourceRoot(context)).pipe(
-//     map(sourceRoot =>
-//       normalizePackagingOptions(options, context.workspaceRoot, sourceRoot)
-//     ),
-//     map(options => 
-//       mergePresetOptions(options)
-//     ),
-//     map(options => 
-//       addDefaultIgnoreOptions(options)
-//     ),
-//     map(options => 
-//       addMissingDefaultOptions(options)
-//     ),
-//     concatMap(async (options) => {
-//       options.afterCopy = [serialHooks([
-//         (buildPath, electronVersion, platform, arch) => {
-//           return writeFileAsync(join(buildPath, 'index.js'), `const Main = require('./dist/apps/${options.name}/main');`);
-//         },
-//         (buildPath, electronVersion, platform, arch) => {
-//           return removeSourceFiles(options, buildPath);
-//         }
-//       ])];
+      const platforms: Platform[] = _createPlatforms(options.platform);
+      const targets: Map<Platform, Map<Arch, string[]>> = _createTargets(platforms, null, options.arch);
+      const baseConfig: Configuration = _createBaseConfig(options, context);
+      const config: Configuration = _createConfigFromOptions(options, baseConfig);
+      const normalizedOptions: CliOptions = _normalizeBuilderOptions(targets, config, rawOptions);
+      const outputPath = await build(normalizedOptions);
 
-//       async function packagerWrapper() {
-//         let result: PackageElectronBuilderOutput;
-//         let outputPath: string | string[];
-//         let success: boolean = true;
+      return { success: true, outputPath };
+    }),
+    catchError(error => {
+      console.error(error);
 
-//         try {
-//           outputPath = await electronPackager(options);
-//         } catch(error) {
-//           success = false;
-//           console.error('Packaging failed:', error);
-//         }
+      return of({ success: false, outputPath: null });
+    })
+  );
+}
 
-//         result = { success, outputPath };
+async function beforeBuild(projectRoot: string, sourcePath: string, appName: string) {
+  await writeFileAsync(join(projectRoot, sourcePath, appName, 'index.js'), `const Main = require('./${appName}/main.js');`);
+}
 
-//         return result;
-//     }
+function _createPlatforms(rawPlatforms: string | string[]): Platform[] {
+  const platforms: Platform[] = [];
 
-//       return await packagerWrapper();
-//     }),
-//     catchError(error => {
-//       console.error(error);
+  if (!rawPlatforms) {
+    const platformMap: Map<string, string> = new Map([['win32', 'windows'], ['darwin', 'mac'], ['linux', 'linux']]); 
 
-//       return of({ success: false, outputPath: null });
-//     })
-//   );
-// }
+    rawPlatforms = platformMap.get(platform());
+  }
 
-// function mergePresetOptions(options: PackageElectronBuilderOptions): PackageElectronBuilderOptions {
-//   // lead preset options file
-//   const externalOptionsPath: string = join(options.dir, options['sourceRoot'], 'app', 'options', 'packager.options.json');
+  if (typeof rawPlatforms === 'string') {
+    rawPlatforms = [rawPlatforms];
+  }
 
-//   if (statSync(externalOptionsPath).isFile()) {
-//     const rawData = readFileSync(externalOptionsPath, 'utf8')
-//     const externalOptions = JSON.parse(stripJsonComments(rawData));
-//     options = Object.assign(normalizeIgnoreOptions(options), normalizeIgnoreOptions(externalOptions));
-//   }
+  if (Array.isArray(rawPlatforms)) {
+    if (rawPlatforms.includes(Platform.WINDOWS.name)) {
+      platforms.push(Platform.WINDOWS);
+    }
 
-//   return options;
-// }
+    if (rawPlatforms.includes(Platform.MAC.name)) {
+      platforms.push(Platform.MAC);
+    }
 
-// function addDefaultIgnoreOptions(options: PackageElectronBuilderOptions): PackageElectronBuilderOptions {
-//   // add ignore options that ignore all the additional projects in the dist folder
-//   // const ignoreExtraProjects: RegExp = new RegExp(`\/dist\/(?!${options.name}$|${options.frontendProject}$).*$`);
-//   const ignoreExtraProjects: Array<RegExp> = getUnrelatedWorkspaceAppsIgnoreList(options);
-  
-//   if (options.ignore) {
-//     if (options.ignore instanceof RegExp) {
-//       ignoreExtraProjects.push(options.ignore);
-//       options.ignore = ignoreExtraProjects;
-//     } else if (Array.isArray(options.ignore)) {
-//       options.ignore.concat(ignoreExtraProjects);
-//     }
-//   } else if (ignoreExtraProjects.length > 0) {
-//     options.ignore = ignoreExtraProjects;
-//   }
+    if (rawPlatforms.includes(Platform.LINUX.name)) {
+      platforms.push(Platform.LINUX);
+    }
+  }
 
-//   return options;
-// }
+  return platforms;
+}
 
-// function getUnrelatedWorkspaceAppsIgnoreList(options: PackageElectronBuilderOptions): Array<RegExp> {
-//   // get regex array of unrelated workspace apps (if exists) to be ignored
-//   let unrelatedAppsPaths: Array<RegExp> = [];
-//   const appsDir: string = 'apps';
-//   const appsPath: string = join(options.dir, appsDir); // assumes that apps is a super set of compiled apps (dist) 
+function _createTargets(platforms: Platform[], type: string, arch: string): Map<Platform, Map<Arch, string[]>> {
+  return createTargets(platforms, null, arch);
+}
 
-//   try {
-//     unrelatedAppsPaths = readdirSync(appsPath, { withFileTypes: true })
-//       .filter(entry => entry.isDirectory())
-//       .filter(entry => entry.name !== options.name && entry.name !== options.frontendProject)
-//       .map(entry => new RegExp(appsDir + '/' + entry.name + '$')) // don't use join here
-//   } catch(error) {
-//     console.error(`${options.name} does not have a valid workspaceRoot. could not generate default ignore list.`);
-//   }
+function _createBaseConfig(options: PackageElectronBuilderOptions, context: BuilderContext): Configuration {
+  const files: Array<FileSet | string> = options.files ?
+   (Array.isArray(options.files) ? options.files : [options.files] ): Array<FileSet | string>()
+  const outputPath = options.prepackageOnly ? 
+    options.outputPath.replace('executables', 'packages') : options.outputPath;
 
-//   return unrelatedAppsPaths;
-// }
+  return {
+    directories: {
+      ...options.directories,
+      output: join(context.workspaceRoot, outputPath)
+    },
+    files: files.concat([
+      {
+          from: resolve(options.sourcePath, options.frontendProject),
+          to: options.frontendProject,
+          filter: ['**/!(*.+(js|css).map)', 'assets']
+      },
+      {
+          from: resolve(options.sourcePath, options.name),
+          to: options.name,
+          filter: ['main.js', 'assets']
+      },
+      {
+          from: resolve(options.sourcePath, options.name),
+          to: '',
+          filter: ['index.js']
+      },      
+      './package.json',
+      '!(**/*.+(js|css).map)',
+    ])
+  };
+}
 
-// function normalizeIgnoreOptions(options: PackageElectronBuilderOptions): PackageElectronBuilderOptions {
-//   // normalize ignore options (if exist) to be of type RegExp | RegExp[]
-//   let normalizedIgnoreOptions: Array<RegExp> = [];
+function _createConfigFromOptions(options: PackageElectronBuilderOptions, baseConfig: Configuration): Configuration {
+  const config = Object.assign({}, options, baseConfig);
+      
+  delete config.name;
+  delete config.frontendProject;
+  delete config.platform;
+  delete config.arch;
+  delete config.root;
+  delete config.prepackageOnly;
+  delete config['sourceRoot'];
+  delete config['$schema'];
+  delete config["publishPolicy"];
+  delete config.sourcePath;
+  delete config.outputPath;
 
-//   if (options.ignore) {
-//     if (typeof options.ignore === 'string') {
-//       options.ignore = new RegExp(options.ignore);
-//     } else if (typeof options.ignore === 'object' && options.ignore instanceof RegExp) {
-//       normalizedIgnoreOptions.push(options.ignore);
-//     } else if (Array.isArray(options.ignore)) {
-//       options.ignore.forEach(option => {
-//         if (typeof option === 'string') {
-//           option = new RegExp(option);
-//         } 
-        
-//         if (option instanceof RegExp) {
-//           normalizedIgnoreOptions.push(option);
-//         }
-//       });
+  return config;
+}
 
-//       options.ignore = normalizedIgnoreOptions;
-//     }
-//   }
+function _normalizeBuilderOptions(targets: Map<Platform, Map<Arch, string[]>>, config: Configuration, rawOptions: JsonObject & PackageElectronBuilderOptions): CliOptions {
+  let normalizedOptions: CliOptions = { config, publish: rawOptions.publishPolicy || null };
 
-//   return options;
-// }
+  if (rawOptions.prepackageOnly) {
+    normalizedOptions.dir = true;
+  } else {
+    normalizedOptions.targets = targets
+  }
 
-// function addMissingDefaultOptions(options: PackageElectronBuilderOptions): PackageElectronBuilderOptions {
-//   //todo: add appVersion
+  return normalizedOptions;
+}
 
-//   // remove unset options (use electron packager default where possible)
-//   Object.keys(options).forEach((key) => (options[key] === '') && delete options[key]);
+function mergePresetOptions(options: PackageElectronBuilderOptions): PackageElectronBuilderOptions {
+  // load preset options file
+  const externalOptionsPath: string = join(options.root, options['sourceRoot'], 'app', 'options', 'maker.options.json');
 
-//   return options;
-// }
+  if (statSync(externalOptionsPath).isFile()) {
+    const rawData = readFileSync(externalOptionsPath, 'utf8')
+    const externalOptions = JSON.parse(stripJsonComments(rawData));
+    options = Object.assign(options, externalOptions);
+  }
 
-// function removeSourceFiles(options: PackageElectronBuilderOptions, buildPath: string): Promise<any> {
-//   // remove source map files
-//   if (options['ignoreSourceMap']) {
-//     if (statSync(join(buildPath, 'dist')).isDirectory()) {
-//       try {
-//         removeSync(join(buildPath, 'dist', '**', '*.js.map'));
-//       } catch (error) {
-//         error('Failed to remove source map files:', error);
-//       }
-//     }
-//   }
+  return options;
+}
 
-//   return new Promise((resolve, reject) => {
-//     // remove source files (./apps directory)
-//     if (statSync(join(buildPath, 'apps')).isDirectory()) {
-//       try {
-//         removeSync(join(buildPath, 'apps'));
-//       } catch (error) {
-//         reject(error);
-//       }
-//     }
+function addMissingDefaultOptions(options: PackageElectronBuilderOptions): PackageElectronBuilderOptions {
+  // remove unset options (use electron builder default values where possible)
+  Object.keys(options).forEach((key) => (options[key] === '') && delete options[key]);
 
-//     resolve();
-//   });
-// }
+  return options;
+}
