+++ src/builders/package/schema.json
@@ -17,37 +17,49 @@
       "description": "Frontend project that needs to access this application.",
       "x-prompt": "What's the name the frontend application you would like to package?"
     },
-    "prune": {
+    "prepackageOnly": {
       "type": "boolean",
-      "default": true,
-      "description": "Remove all unneeded dependencies."
+      "description": "Only prepackage the aplication.",
+      "default": false
     },
-    "asar": {
-      "type": "boolean",
-      "default": false,
-      "description": "Pack app inside an asar file."
+    "sourcePath": {
+      "type": "string",
+      "description": "The full path for the compiled source directory, relative to the current workspace. If this parameter is not supplied, the 'sourcePath' path from angular.json will be used.",
+      "default": "dist/apps"
     },
-    "overwrite": {
-      "type": "boolean",
-      "default": false,
-      "description": "Overwrite previous packaged applications."
+    "outputPath": {
+      "type": "string",
+      "description": "The full path for the new output directory, relative to the current workspace. If this parameter is not supplied, the 'outputPath' path from angular.json will be used."
+    },
+    "platform": {
+      "oneOf": [
+        {
+          "type": "string",
+          "enum": ["windows", "mac", "linux"]
+        },
+        {
+          "type": "array",
+          "uniqueItems": true,
+          "items": {
+            "type": "string",
+            "enum": ["windows", "mac", "linux"]
+          }
+        }
+      ],
+      "default": null,
+      "description": "Target platform/s."
     },
     "arch": {
       "type": "string",
-      "default": "",
-      "enum": ["ia32", "x64", "armv7l", "arm64", "mips64el", "all"],
+      "default": null,
+      "enum": ["x64", "ia32", "armv7l", "arm64"],
       "description": "Package architecture."
     },
-    "platform": {
+    "publishPolicy": {
       "type": "string",
-      "default": "",
-      "enum": ["darwin", "linux", "mas", "win32", "all"],
-      "description": "Target platform."
-    },
-    "ignore": {
-      "type": "string",
-      "default": "(libs|tools)",
-      "description": "Files and directories to ignore (Regex)."
+      "default": null,
+      "enum": ["onTag", "onTagOrDraft", "always", "never"],
+      "description": "Package publish policy."
     }
   },
   "required": ["name", "frontendProject"]
